move file of confirmation dialog
